split callbacks.py into a directory
