+++ emote/extra/onnx_exporter.py
@@ -13,7 +13,7 @@ import torch
 from google.protobuf import text_format
 
 from emote.callback import Callback
-from emote.callbacks import LoggingMixin
+from emote.callbacks.logging import LoggingMixin
 from emote.extra.crud_storage import CRUDStorage, StorageItem, StorageItemHandle
 from emote.proxies import AgentProxy
 from emote.utils.spaces import MDPSpace
