+++ emote/memory/memory.py
@@ -18,7 +18,7 @@ from typing import Callable, Dict, List, Mapping, Optional, Tuple, Union
 from torch.utils.tensorboard import SummaryWriter
 
 from emote.callback import Callback
-from emote.callbacks import LoggingMixin
+from emote.callbacks.logging import LoggingMixin
 from emote.extra.onnx_exporter import OnnxExporter
 from emote.trainer import TrainingShutdownException
