+++ experiments/gym/train_lunar_lander.py
@@ -15,12 +15,8 @@ from torch.optim import Adam
 from torch.utils.tensorboard import SummaryWriter
 
 from emote import Trainer
-from emote.callbacks import (
-    BackPropStepsTerminator,
-    LoggingMixin,
-    TensorboardLogger,
-    WBLogger,
-)
+from emote.callbacks.generic import BackPropStepsTerminator
+from emote.callbacks.logging import LoggingMixin, TensorboardLogger, WBLogger
 from emote.memory import MemoryLoader, TableMemoryProxy
 from emote.memory.builder import DictObsNStepTable
 from emote.nn import GaussianPolicyHead
