+++ experiments/gym/wandb_sweep_lunar_lander.py
@@ -1,5 +1,4 @@
 import argparse
-import time
 
 from functools import partial
 
@@ -15,7 +14,8 @@ from torch import nn
 from torch.optim import Adam
 
 from emote import Trainer
-from emote.callbacks import BackPropStepsTerminator, WBLogger
+from emote.callbacks.generic import BackPropStepsTerminator
+from emote.callbacks.logging import WBLogger
 from emote.memory import MemoryLoader, TableMemoryProxy
 from emote.memory.builder import DictObsNStepTable
 from emote.nn import GaussianPolicyHead
