+++ tests/gym/collector.py
@@ -9,7 +9,7 @@ from collections import deque
 from tests.gym.dict_gym_wrapper import DictGymWrapper
 
 from emote.callback import Callback
-from emote.callbacks import LoggingMixin
+from emote.callbacks.logging import LoggingMixin
 from emote.proxies import AgentProxy, MemoryProxy
