+++ tests/test_wandb.py
@@ -1,8 +1,10 @@
 import os
 
+import pytest
 import wandb
 
-from emote.callbacks import Callback, LoggingMixin, WBLogger
+from emote.callback import Callback
+from emote.callbacks.logging import LoggingMixin, WBLogger
 from emote.trainer import Trainer
 
 
@@ -24,6 +26,17 @@ class DummyLoader:
     def __iter__(self):
         for _ in range(N):
             yield {"batch_size": 0}
+
+
+def test_raises_help_if_wandb_not_installed(hide_pkg):
+    hide_pkg("wandb")
+
+    with pytest.raises(ImportError) as ex:
+        WBLogger([], {}, log_interval=1)
+
+    assert ex.value.msg == "enable the optional `wandb` future to use the WBLogger"
+    assert isinstance(ex.value.__cause__, ImportError)
+    assert ex.value.__cause__.msg == "No module named 'wandb'"
 
 
 def test_logging():
